Finish Lecture 44 - Clean up raycasting
The raycasting code is now centralized in a way that you don't need to change three different scripts every time a layer is added. The Utility script is removed for this very reason - the Layer enum has been removed in favour of using integers instead, and an editor script was added to simplify this transition.