Finish Lecture 55 - Make enemies shoot at player
Each enemy now spawns a projectile prefab every frame that the player is within their attackRadius and sends it towards the player.

NOTE: Not yet fully implemented, as they should not fire every single frame, of course. They also fire towards the feet of the player, but we'll deal with that later.