Add an early implementation of the combat controller
A CombatController script now keeps track of combat state. When a player attacks or is attacked the controller is notified, so that it can change state properly.

Additionally, auto attacks are now implemented. When the player clicks an enemy it is now targeted, and if it is close enough the player will attack it automatically.